requirements.txt
numpy
scikit-learn
pillow
p_tqdm
